server-log-analytics/requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Database - ClickHouse
clickhouse-driver==0.2.6
clickhouse-connect==0.6.23

# Alternative: BigQuery
# google-cloud-bigquery==3.14.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
dask==2023.12.1  # For large datasets

# Log Parsing
python-dateutil==2.8.2
user-agents==2.2.0  # User-agent parsing
maxminddb-geolite2==2018.701  # GeoIP

# Analytics & ML
scikit-learn==1.3.2
scipy==1.11.4

# Dashboards
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Monitoring (optional)
# grafana-client==3.5.0

# CLI
click==8.1.7
rich==13.7.0

# Utilities
pyyaml==6.0.1
validators==0.22.0

# Testing
pytest==7.4.3

# Development
black==23.12.1
ipython==8.18.1
